+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/BehaviourContext.kt
@@ -10,6 +10,7 @@ import kotlinx.coroutines.flow.filter
 
 typealias BehaviourContextReceiver<T> = suspend BehaviourContext.() -> T
 typealias BehaviourContextAndTypeReceiver<T, I> = suspend BehaviourContext.(I) -> T
+typealias BehaviourContextAndTwoTypesReceiver<T, I1, I2> = suspend BehaviourContext.(I1, I2) -> T
 
 /**
  * This class contains all necessary tools for work with bots and especially for [buildBehaviour]