+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/triggers_handling/CommandHandling.kt
@@ -1,10 +1,10 @@
 package dev.inmo.tgbotapi.extensions.behaviour_builder.triggers_handling
 
-import dev.inmo.tgbotapi.extensions.behaviour_builder.BehaviourContext
-import dev.inmo.tgbotapi.extensions.behaviour_builder.BehaviourContextAndTypeReceiver
+import dev.inmo.tgbotapi.extensions.behaviour_builder.*
 import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.marker_factories.ByChatMessageMarkerFactory
 import dev.inmo.tgbotapi.extensions.behaviour_builder.utils.marker_factories.MarkerFactory
 import dev.inmo.tgbotapi.extensions.utils.asBotCommandTextSource
+import dev.inmo.tgbotapi.extensions.utils.extensions.parseCommandsWithParams
 import dev.inmo.tgbotapi.types.message.abstracts.CommonMessage
 import dev.inmo.tgbotapi.types.message.content.TextContent
 import kotlinx.coroutines.Job
@@ -33,6 +33,7 @@ suspend fun BehaviourContext.command(
     markerFactory,
     scenarioReceiver
 )
+
 suspend fun BehaviourContext.command(
     command: String,
     requireOnlyCommandInMessage: Boolean = true,
@@ -59,3 +60,53 @@ suspend inline fun BehaviourContext.onCommand(
     markerFactory: MarkerFactory<in CommonMessage<TextContent>, Any> = ByChatMessageMarkerFactory,
     noinline scenarioReceiver: BehaviourContextAndTypeReceiver<Unit, CommonMessage<TextContent>>
 ): Job = onCommand(command.toRegex(), requireOnlyCommandInMessage, includeFilterByChatInBehaviourSubContext, additionalFilter, markerFactory, scenarioReceiver)
+
+suspend fun BehaviourContext.commandWithArgs(
+    commandRegex: Regex,
+    includeFilterByChatInBehaviourSubContext: Boolean = true,
+    additionalFilter: CommonMessageFilter<TextContent>? = null,
+    markerFactory: MarkerFactory<in CommonMessage<TextContent>, Any> = ByChatMessageMarkerFactory,
+    scenarioReceiver: BehaviourContextAndTwoTypesReceiver<Unit, CommonMessage<TextContent>, Array<String>>
+) = command(
+    commandRegex,
+    requireOnlyCommandInMessage = false,
+    includeFilterByChatInBehaviourSubContext = includeFilterByChatInBehaviourSubContext,
+    additionalFilter = additionalFilter,
+    markerFactory = markerFactory
+) {
+    val args = it.parseCommandsWithParams().let { commandsWithArgs ->
+        val key = commandsWithArgs.keys.firstOrNull { it.matches(commandRegex) } ?: return@let null
+        commandsWithArgs[key]
+    } ?: emptyArray()
+    scenarioReceiver(it, args)
+}
+
+suspend fun BehaviourContext.commandWithArgs(
+    command: String,
+    includeFilterByChatInBehaviourSubContext: Boolean = true,
+    additionalFilter: CommonMessageFilter<TextContent>? = null,
+    markerFactory: MarkerFactory<in CommonMessage<TextContent>, Any> = ByChatMessageMarkerFactory,
+    scenarioReceiver: BehaviourContextAndTwoTypesReceiver<Unit, CommonMessage<TextContent>, Array<String>>
+) = commandWithArgs(
+    command.toRegex(),
+    includeFilterByChatInBehaviourSubContext = includeFilterByChatInBehaviourSubContext,
+    additionalFilter = additionalFilter,
+    markerFactory = markerFactory,
+    scenarioReceiver = scenarioReceiver
+)
+
+suspend inline fun BehaviourContext.onCommandWithArgs(
+    commandRegex: Regex,
+    includeFilterByChatInBehaviourSubContext: Boolean = true,
+    noinline additionalFilter: CommonMessageFilter<TextContent>? = null,
+    markerFactory: MarkerFactory<in CommonMessage<TextContent>, Any> = ByChatMessageMarkerFactory,
+    noinline scenarioReceiver: BehaviourContextAndTwoTypesReceiver<Unit, CommonMessage<TextContent>, Array<String>>
+): Job = commandWithArgs(commandRegex, includeFilterByChatInBehaviourSubContext, additionalFilter, markerFactory, scenarioReceiver)
+
+suspend inline fun BehaviourContext.onCommandWithArgs(
+    command: String,
+    includeFilterByChatInBehaviourSubContext: Boolean = true,
+    noinline additionalFilter: CommonMessageFilter<TextContent>? = null,
+    markerFactory: MarkerFactory<in CommonMessage<TextContent>, Any> = ByChatMessageMarkerFactory,
+    noinline scenarioReceiver: BehaviourContextAndTwoTypesReceiver<Unit, CommonMessage<TextContent>, Array<String>>
+): Job = onCommandWithArgs(command.toRegex(), includeFilterByChatInBehaviourSubContext, additionalFilter, markerFactory, scenarioReceiver)